ResourceMap: Make map determine what markers are in view
The process is that, on map update an event is emitted to the App
component which then performs calculations to determine which markers
are within view to apply additional updates to. This is not ideal
because instead, the map should tell the parent component what markers
in view. Therefore, make the map determine what markers are in view and
propagate those markers up the event chain.